// web/src/generated/api/proto/physics/vqe.ts
export interface MoleculeConfig {
    /**
     * @generated from protobuf field: string name = 1
     */
    name: string; // e.g. "H2", "LiH", "HeH+"
    /**
     * @generated from protobuf field: repeated qubit_engine.physics.Atom atoms = 2
     */
    atoms: Atom[]; // Atom positions
    /**
     * @generated from protobuf field: int32 charge = 3
     */
    charge: number; // Molecular charge
    /**
     * @generated from protobuf field: int32 multiplicity = 4
     */
    multiplicity: number; // Spin multiplicity (1=singlet, 2=doublet, etc.)
    /**
     * @generated from protobuf field: string basis_set = 5
     */
    basisSet: string; // e.g. "sto-3g", "6-31g"
}
/**
 * @generated from protobuf message qubit_engine.physics.Atom
 */
export interface Atom {
    /**
     * @generated from protobuf field: string element = 1
     */
    element: string; // e.g. "H", "Li", "He"
    /**
     * @generated from protobuf field: double x = 2
     */
    x: number; // Position in Angstroms
    /**
     * @generated from protobuf field: double y = 3
     */
    y: number;
    /**
     * @generated from protobuf field: double z = 4
     */
    z: number;
}
// ------------------------------------------------------------------
// Hamiltonian Representation (Pauli String Sum)
// H = Σ c_i * P_i where P_i is a Pauli string (tensor product of I, X, Y, Z)
// ------------------------------------------------------------------

/**
 * @generated from protobuf message qubit_engine.physics.Hamiltonian
 */
export interface Hamiltonian {
    /**
     * @generated from protobuf field: string molecule_name = 1
     */
    moleculeName: string;
    /**
     * @generated from protobuf field: int32 num_qubits = 2
     */
    numQubits: number; // Number of qubits needed
    /**
     * @generated from protobuf field: repeated qubit_engine.physics.PauliTerm terms = 3
     */
    terms: PauliTerm[]; // Sum of Pauli terms
    /**
     * @generated from protobuf field: double nuclear_repulsion = 4
     */
    nuclearRepulsion: number; // Nuclear repulsion energy (constant offset)
}
/**
 * @generated from protobuf message qubit_engine.physics.PauliTerm
 */
export interface PauliTerm {
    /**
     * @generated from protobuf field: double coefficient = 1
     */
    coefficient: number; // Real coefficient
    /**
     * @generated from protobuf field: repeated qubit_engine.physics.PauliOperator operators = 2
     */
    operators: PauliOperator[];
}
/**
 * @generated from protobuf message qubit_engine.physics.PauliOperator
 */
export interface PauliOperator {
    /**
     * @generated from protobuf field: int32 qubit = 1
     */
    qubit: number;
    /**
     * @generated from protobuf field: qubit_engine.physics.PauliType type = 2
     */
    type: PauliType;
}
/**
 * @generated from protobuf message qubit_engine.physics.VQERequest
 */
export interface VQERequest {
    /**
     * @generated from protobuf oneof: target
     */
    target: {
        oneofKind: "molecule";
        /**
         * @generated from protobuf field: qubit_engine.physics.MoleculeConfig molecule = 1
         */
        molecule: MoleculeConfig;
    } | {
        oneofKind: "hamiltonian";
        /**
         * @generated from protobuf field: qubit_engine.physics.Hamiltonian hamiltonian = 2
         */
        hamiltonian: Hamiltonian;
    } | {
        oneofKind: undefined;
    };
    /**
     * @generated from protobuf field: qubit_engine.physics.AnsatzType ansatz = 3
     */
    ansatz: AnsatzType;
    /**
     * @generated from protobuf field: qubit_engine.physics.OptimizerType optimizer = 4
     */
    optimizer: OptimizerType;
    /**
     * @generated from protobuf field: int32 max_iterations = 5
     */
    maxIterations: number;
    /**
     * @generated from protobuf field: double convergence_threshold = 6
     */
    convergenceThreshold: number;
    /**
     * @generated from protobuf field: repeated double initial_parameters = 7
     */
    initialParameters: number[]; // Optional starting point
    /**
     * @generated from protobuf field: int32 shots_per_evaluation = 8
     */
    shotsPerEvaluation: number;
}
// ...
export enum AnsatzType {
    /**
     * Unitary Coupled Cluster Singles and Doubles
     *
     * @generated from protobuf enum value: ANSATZ_UCCSD = 0;
     */
    ANSATZ_UCCSD = 0,
    /**
     * Hardware efficient ansatz
     *
     * @generated from protobuf enum value: ANSATZ_HARDWARE_EFFICIENT = 1;
     */
    ANSATZ_HARDWARE_EFFICIENT = 1,
    /**
     * Simple RY rotation layers
     *
     * @generated from protobuf enum value: ANSATZ_RY = 2;
     */
    ANSATZ_RY = 2
}
/**
 * @generated from protobuf enum qubit_engine.physics.OptimizerType
 */
export enum OptimizerType {
    /**
     * @generated from protobuf enum value: OPTIMIZER_COBYLA = 0;
     */
    OPTIMIZER_COBYLA = 0,
    /**
     * @generated from protobuf enum value: OPTIMIZER_SPSA = 1;
     */
    OPTIMIZER_SPSA = 1,
    /**
     * @generated from protobuf enum value: OPTIMIZER_ADAM = 2;
     */
    OPTIMIZER_ADAM = 2,
    /**
     * @generated from protobuf enum value: OPTIMIZER_GRADIENT_DESCENT = 3;
     */
    OPTIMIZER_GRADIENT_DESCENT = 3
}
// ...
class MoleculeConfig$Type extends MessageType<MoleculeConfig> {
    constructor() {
        super("qubit_engine.physics.MoleculeConfig", [
            { no: 1, name: "name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "atoms", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => Atom },
            { no: 3, name: "charge", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 4, name: "multiplicity", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 5, name: "basis_set", kind: "scalar", T: 9 /*ScalarType.STRING*/ }
        ]);
    }
    create(value?: PartialMessage<MoleculeConfig>): MoleculeConfig {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.name = "";
        message.atoms = [];
        message.charge = 0;
        message.multiplicity = 0;
        message.basisSet = "";
        if (value !== undefined)
            reflectionMergePartial<MoleculeConfig>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: MoleculeConfig): MoleculeConfig {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string name */ 1:
                    message.name = reader.string();
                    break;
                case /* repeated qubit_engine.physics.Atom atoms */ 2:
                    message.atoms.push(Atom.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* int32 charge */ 3:
                    message.charge = reader.int32();
                    break;
                case /* int32 multiplicity */ 4:
                    message.multiplicity = reader.int32();
                    break;
                case /* string basis_set */ 5:
                    message.basisSet = reader.string();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: MoleculeConfig, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string name = 1; */
        if (message.name !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.name);
        /* repeated qubit_engine.physics.Atom atoms = 2; */
        for (let i = 0; i < message.atoms.length; i++)
            Atom.internalBinaryWrite(message.atoms[i], writer.tag(2, WireType.LengthDelimited).fork(), options).join();
        /* int32 charge = 3; */
        if (message.charge !== 0)
            writer.tag(3, WireType.Varint).int32(message.charge);
        /* int32 multiplicity = 4; */
        if (message.multiplicity !== 0)
            writer.tag(4, WireType.Varint).int32(message.multiplicity);
        /* string basis_set = 5; */
        if (message.basisSet !== "")
            writer.tag(5, WireType.LengthDelimited).string(message.basisSet);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message qubit_engine.physics.MoleculeConfig
 */
export const MoleculeConfig = new MoleculeConfig$Type();
// ...
export const Atom = new Atom$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Hamiltonian$Type extends MessageType<Hamiltonian> {
    constructor() {
        super("qubit_engine.physics.Hamiltonian", [
            { no: 1, name: "molecule_name", kind: "scalar", T: 9 /*ScalarType.STRING*/ },
            { no: 2, name: "num_qubits", kind: "scalar", T: 5 /*ScalarType.INT32*/ },
            { no: 3, name: "terms", kind: "message", repeat: 2 /*RepeatType.UNPACKED*/, T: () => PauliTerm },
            { no: 4, name: "nuclear_repulsion", kind: "scalar", T: 1 /*ScalarType.DOUBLE*/ }
        ]);
    }
    create(value?: PartialMessage<Hamiltonian>): Hamiltonian {
        const message = globalThis.Object.create((this.messagePrototype!));
        message.moleculeName = "";
        message.numQubits = 0;
        message.terms = [];
        message.nuclearRepulsion = 0;
        if (value !== undefined)
            reflectionMergePartial<Hamiltonian>(this, message, value);
        return message;
    }
    internalBinaryRead(reader: IBinaryReader, length: number, options: BinaryReadOptions, target?: Hamiltonian): Hamiltonian {
        let message = target ?? this.create(), end = reader.pos + length;
        while (reader.pos < end) {
            let [fieldNo, wireType] = reader.tag();
            switch (fieldNo) {
                case /* string molecule_name */ 1:
                    message.moleculeName = reader.string();
                    break;
                case /* int32 num_qubits */ 2:
                    message.numQubits = reader.int32();
                    break;
                case /* repeated qubit_engine.physics.PauliTerm terms */ 3:
                    message.terms.push(PauliTerm.internalBinaryRead(reader, reader.uint32(), options));
                    break;
                case /* double nuclear_repulsion */ 4:
                    message.nuclearRepulsion = reader.double();
                    break;
                default:
                    let u = options.readUnknownField;
                    if (u === "throw")
                        throw new globalThis.Error(`Unknown field ${fieldNo} (wire type ${wireType}) for ${this.typeName}`);
                    let d = reader.skip(wireType);
                    if (u !== false)
                        (u === true ? UnknownFieldHandler.onRead : u)(this.typeName, message, fieldNo, wireType, d);
            }
        }
        return message;
    }
    internalBinaryWrite(message: Hamiltonian, writer: IBinaryWriter, options: BinaryWriteOptions): IBinaryWriter {
        /* string molecule_name = 1; */
        if (message.moleculeName !== "")
            writer.tag(1, WireType.LengthDelimited).string(message.moleculeName);
        /* int32 num_qubits = 2; */
        if (message.numQubits !== 0)
            writer.tag(2, WireType.Varint).int32(message.numQubits);
        /* repeated qubit_engine.physics.PauliTerm terms = 3; */
        for (let i = 0; i < message.terms.length; i++)
            PauliTerm.internalBinaryWrite(message.terms[i], writer.tag(3, WireType.LengthDelimited).fork(), options).join();
        /* double nuclear_repulsion = 4; */
        if (message.nuclearRepulsion !== 0)
            writer.tag(4, WireType.Bit64).double(message.nuclearRepulsion);
        let u = options.writeUnknownFields;
        if (u !== false)
            (u == true ? UnknownFieldHandler.onWrite : u)(this.typeName, message, writer);
        return writer;
    }
}
/**
 * @generated MessageType for protobuf message qubit_engine.physics.Hamiltonian
 */
export const Hamiltonian = new Hamiltonian$Type();
// ...
export const PauliTerm = new PauliTerm$Type();
// ...
export const VQERequest = new VQERequest$Type();